reskin UnlockAccount view (#362)
* reskin UnlockAccount

* wrong parenthesis in other forms

* remove <br />s
