+++ extension/src/popup/components/mnemonicPhrase/ConfirmMnemonicPhrase/index.tsx
@@ -110,6 +110,7 @@ export const ConfirmMnemonicPhrase = ({
               <FormRow>
                 <Button
                   fullWidth
+                  type="submit"
                   disabled={!dirty && !!displaySelectedWords().length}
                   isLoading={isSubmitting}
                 >
